pyproject.toml
[project]
name = "md4-mae-min"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "absl-py>=2.1.0",
    "clu>=0.0.12",
    "etils[epath]>=1.13.0",
    "flax==0.12.2",
    "huggingface-hub>=0.33.2",
    "h5py>=3.11.0",
    "jax[cuda]==0.8.2; sys_platform == 'linux'",
    "jax==0.8.2; sys_platform == 'macOS'",
    "matplotlib>=3.10.3",
    "ml-collections>=1.1.0",
    "optax==0.2.6",
    "orbax-checkpoint==0.11.30",
    "rdkit>=2025.3.3",
    "scikit-learn>=1.8.0",
    "tensorflow==2.19.0",
    "tensorboard==2.19.0",
    "tqdm>=4.66.4",
    "wandb==0.23.1",
]
